requirements.txt
streamlit
streamlit-authenticator
google-cloud-firestore
python-dotenv
pandas
altair
toml
